+++ smods.lua
@@ -13,7 +13,7 @@ if SMODS.current_mod then
     local configSuccess, config = pcall(require, "debugplus-config")
     
     if not configSuccess then
-        error("DebugPlus modules not successfully initialized.\nMake sure your DebugPlus folder is not nested (there should be a bunch of files in the DebugPlus folder and not just another folder).")
+        error("DebugPlus modules not successfully initialized.\nMake sure your DebugPlus folder is not nested (there should be a bunch of files in the DebugPlus folder and not just another folder).\n\n" .. (config or "No further info."))
     end
     SMODS.current_mod.config_tab = config.generateConfigTab
     config.SMODSLoaded = true